requirements.txt
Django==5.1.14
groq==0.4.2
python-dotenv==1.0.0
requests==2.31.0

